py-pylint: updated to 3.0.3
What's new in Pylint 3.0.3?

False Positives Fixed

Fixed false positive for unnecessary-lambda when the call has keyword arguments but not the lambda.

Fixed incorrect suggestion for shallow copy in unnecessary-comprehension

Example of the suggestion:
#pylint: disable=missing-module-docstring
a = [1, 2, 3]
b = [x for x in a]
b[0] = 0
print(a) # [1, 2, 3]

After changing b = [x for x in a] to b = a based on the suggestion, the script now prints [0, 2, 3]. The correct suggestion should be use list(a) to preserve the original behavior.

Fix false positives for undefined-variable and unused-argument for
classes and functions using Python 3.12 generic type syntax.

Fixed pointless-string-statement false positive for docstrings
on Python 3.12 type aliases.

Fix false positive for invalid-exception-operation when concatenating tuples
of exception types.

Other Bug Fixes

Fix a bug where pylint was unable to walk recursively through a directory if the
directory has an __init__.py file.