gh: update to 2.40.1
Multiple Account Support

With the v2.40.0 release, multi-account support is introduced to gh for
GitHub.com and GitHub Enterprise and it is only the beginning!

This release focuses on several specific use cases:

-   adding multiple accounts for GitHub.com and GitHub Enterprise with gh auth
    login
-   switching manually between accounts for gh and git use with gh auth switch
-   seeing multiple accounts with gh auth status
-   logging out of accounts with gh auth logout

Future enhancements such as automatic account switching based on context and
more git automatic git configuration will be planned based on community
feedback.

For much more detail about the specific of multiple account support in this
release as well as a discussion of the sharp edges, please see this doc.

Please provide feedback in our release discussion.