py-hypothesis: updated to 6.92.5
6.92.5

This patch registers explicit strategies for a handful of builtin types, motivated by improved introspection in PyPy 7.3.14 triggering existing internal warnings. Thanks to Carl Friedrich Bolz-Tereick for helping us work out what changed!

6.92.4

This patch fixes an error when writing observability reports without a pre-existing .hypothesis directory.

6.92.3

This patch adds a new environment variable HYPOTHESIS_EXPERIMENTAL_OBSERVABILITY_NOCOVER, which turns on observability data collection without collecting code coverage data, which may be faster on Python 3.11 and earlier.

6.92.2

This patch updates our vendored list of top-level domains, which is used by the provisional domains() strategy.

6.92.1

This patch fixes a bug introduced in version 6.92.0, where using the data() strategy would fail to draw a dataclass() with a defaultdict field. This was due to a bug in the standard library which was fixed in 3.12, so we’ve vendored the fix.