security/heimdal, security/mit-krb5: Do krb5-config in cross-destdir.
Although this is an executable program, it's a very simple shell
script which can be run in the build environment to describe the
target environment.  So qualify paths to it by ${_CROSS_DESTDIR:U},
where it lives during cross-builds.

No change to native builds because ${_CROSS_DESTDIR:U} is empty in
them.