gtk4: build workaround for NetBSD with gcc 12
NetBSD ld.elf_so dumps core with a shlib built with gcc 12 and
both -z,relro and -z,now. c.f., PR lib/57792

gtk4 adds both of these flags by default, and fails on either a
system with a base HAVE_GCC=12, or with GCC_REQD=12.

With the patch, it can be caused to fail again by setting
PKGSRC_USE_RELRO=full

Thank you to RVP for the analysis